group settings and ui

+++ backend/controllers/groupController.js
@@ -2,6 +2,7 @@ const Group = require('../models/Group')
 const User = require('../models/User')
 const jwt = require('jsonwebtoken')
 const mongoose = require('mongoose')
+const validator = require('validator')
 
 //create new group
 const createGroup = async (req, res) => {
@@ -13,8 +14,9 @@ const createGroup = async (req, res) => {
 
     try {      
         const group = await Group.createGroup(name, user, sectionID)
-        res.status(200).json(await group.populate('createdByID'))
+        res.status(200).json(await group.getSafeData())
     } catch(error) {
+        console.log(error.message)
         res.status(400).json({error: error.message})
     }
 }
@@ -23,7 +25,7 @@ const createGroup = async (req, res) => {
 const addUser = async (req, res) => {
     const { id } = req.params
     const { user_id } = req.body
-
+    
     const group = await Group.findById(id)
     if (!group) {
         return res.status(404).json({ error: "Group not found." });
@@ -36,12 +38,17 @@ const addUser = async (req, res) => {
 
     const isUserRequest = group.requestID.some(requestUser => requestUser.equals(user_id));
     if (isUserRequest) {
-        return res.status(400).json({ error: "User is already in the pending list." });
+        return res.status(400).json({ error: "User is already in the request list." });
     }
 
-    const isUserMember = group.membersID.some(member => member.equals(user_id))
+    const isUserMember = await group.isMember(user_id)
     if (isUserMember) {
         return res.status(400).json({ error: "User is already a member." });
+    }
+
+    const canManage = group.canManage(req.user._id)
+    if (!canManage) {
+        return res.status(400).json({ error: "User has no privileges." });
     }
 
     try {
@@ -58,7 +65,7 @@ const addUser = async (req, res) => {
 //accept invite
 const acceptGroup = async (req, res) => {
     const { id } = req.params
-    const { user_id } = req.body
+    const user_id = req.user._id
 
     const group = await Group.findById(id)
     if (!group) {
@@ -70,14 +77,15 @@ const acceptGroup = async (req, res) => {
         return res.status(400).json({ error: "User is not in the pending list." });
     }
 
-    const isUserMember = group.membersID.some(member => member.equals(user_id))
+    const isUserMember = await group.isMember(user_id)
     if (isUserMember) {
         return res.status(400).json({ error: "User is already a member." });
     }
 
     try {
-        group.addMember(user_id)
-        res.status(200).json(await Group.find({_id: id }))
+        await group.addMember(user_id)
+        console.log(group.membersID)
+        res.status(200).json(await group.getSafeData())
     } catch (error) {
         res.status(400).json({error: error.message})
     }
@@ -86,7 +94,7 @@ const acceptGroup = async (req, res) => {
 //reject group
 const rejectGroup = async (req, res) => {
     const { id } = req.params
-    const { user_id } = req.body
+    const user_id = req.user._id
 
     const group = await Group.findById(id)
     if (!group) {
@@ -98,14 +106,14 @@ const rejectGroup = async (req, res) => {
         return res.status(400).json({ error: "User is not in the pending list." });
     }
 
-    const isUserMember = group.membersID.some(member => member.equals(user_id))
+    const isUserMember = await group.isMember(user_id)
     if (isUserMember) {
         return res.status(400).json({ error: "User is already a member." });
     }
 
     const isUserOwner = group.createdByID.equals(user_id)
     if (isUserOwner) {
-        return res.status(400).json({ error: "User is already a member." });
+        return res.status(400).json({ error: "User is the owner." });
     }
 
     try {
@@ -113,7 +121,7 @@ const rejectGroup = async (req, res) => {
             { _id: id},
             { $pull: { pendingID: user_id}}
         )
-        res.status(200).json(await Group.find({_id: id }))
+        res.status(200).json(true)
     } catch (error) {
         res.status(400).json({error: error.message})
     }
@@ -126,7 +134,8 @@ const getAllGroups = async (req, res) => {
     const groups = await Group.find({
         $or: [
           { membersID: { $all: [user_id] } },
-          { createdByID: user_id }
+          { createdByID: user_id },
+          { requestID: { $all: [user_id] } },
         ]
       })
       .sort({ "createdAt": 1 })
@@ -169,7 +178,7 @@ const removeMember = async (req, res) => {
         return res.status(404).json({ error: "Group not found." });
     }
 
-    const isUserMember = group.membersID.some(member => member.equals(user_id))
+    const isUserMember = await group.isMember(user_id)
     if (!isUserMember) {
         return res.status(400).json({ error: "User is not a member." });
     }
@@ -179,9 +188,14 @@ const removeMember = async (req, res) => {
         return res.status(400).json({ error: "User is the owner." });
     }
 
+    const canManage = group.canManage(req.user._id)
+    if (!canManage) {
+        return res.status(400).json({ error: "User has no privileges." });
+    }
+
     try {
-        group.removeMember(user_id)
-        res.status(200).json(await (await Group.findById(id)).getSafeData())
+        await group.removeMember(user_id)
+        res.status(200).json(await group.getSafeData())
     } catch (error) {
         res.status(400).json({error: error.message})
     }
@@ -200,11 +214,16 @@ const deleteGroup = async (req, res) => {
     if (!group) {
         return res.status(404).json({ error: "Group not found." });
     } else {
-       await Group.findOneAndDelete({
+        if (!group.createdByID.equals(user_id)) {
+            return res.status(400).json({ error: "User does not have privileges." });
+        }
+
+        await Group.findOneAndDelete({
             _id: id,
             createdByID: user_id,
         })
         await Group.model('ChatChannel').findOneAndDelete({ _id: group.chatChannelID })
+        await Group.model('TaskSection').deleteMany({ group_id: group._id })
     }
 
     res.status(200).json(group)
@@ -234,7 +253,7 @@ const searchGroup = async (req, res) => {
 
 const joinGroup = async (req, res) => {
     const { id } = req.params
-    const { user_id } = req.body
+    const user_id = req.user._id
 
     const group = await Group.findById(id)
     if (!group) {
@@ -251,7 +270,7 @@ const joinGroup = async (req, res) => {
         return res.status(400).json({ error: "User is already in the pending list." });
     }
 
-    const isUserMember = group.membersID.some(member => member.equals(user_id))
+    const isUserMember = await group.isMember(user_id)
     if (isUserMember) {
         return res.status(400).json({ error: "User is already a member." });
     }
@@ -266,9 +285,78 @@ const joinGroup = async (req, res) => {
             { _id: id},
             {$push: { requestID: user_id}}
         )
+        res.status(200).json(true)
+    } catch (error) {
+        console.log(error.message)
+        res.status(400).json({error: error.message})
+    }
+}
+
+const revokeInvite = async (req, res) => {
+    const { id } = req.params
+    const { user_id } = req.body
+
+    const group = await Group.findById(id)
+    if (!group) {
+        return res.status(404).json({ error: "Group not found." });
+    }
+
+    const isUserPending = group.pendingID.some(pendingUser => pendingUser.equals(user_id));
+    if (!isUserPending) {
+        return res.status(400).json({ error: "User is not in the pending list." });
+    }
+
+    const canManage = group.canManage(req.user._id)
+    if (canManage) {
+        await Group.updateOne(
+            { _id: id},
+            { $pull: { pendingID: user_id}}
+        )
         res.status(200).json(await (await Group.findById(id)).getSafeData())
+    } else {
+        return res.status(400).json({ error: "User has no privileges." });
+    }
+}
+
+const cancelRequest = async (req, res) => {
+    const { id } = req.params
+    const user_id = req.user._id
+
+    const group = await Group.findById(id)
+    if (!group) {
+        return res.status(404).json({ error: "Group not found." });
+    }
+    
+    const isUserRequest = group.requestID.some(requestUser => requestUser.equals(user_id));
+    if (!isUserRequest) {
+        return res.status(400).json({ error: "User is not in the request list." });
+    }
+
+    try {
+        await Group.updateOne(
+            { _id: id},
+            { $pull: { requestID: user_id}}
+        )
+        res.status(200).json(await group.getSafeData())
     } catch (error) {
         res.status(400).json({error: error.message})
+    }
+}
+
+const leaveGroup = async (req, res) => {
+    const { id } = req.params
+    const user_id = req.user._id
+
+    const group = await Group.findById(id)
+    if (!group) {
+        return res.status(404).json({ error: "Group not found." });
+    }
+    
+    if (await group.isMember(user_id) && !(await group.isOwner(user_id))) {
+        await group.removeMember(user_id)
+        res.status(200).json(true)
+    } else {
+        return res.status(400).json({ error: "User is not a member or the owner." });
     }
 }
 
@@ -307,14 +395,19 @@ const acceptUser = async (req, res) => {
         return res.status(400).json({ error: "User is not in the pending list." });
     }
 
-    const isUserMember = group.membersID.some(member => member.equals(user_id))
+    const isUserMember = await group.isMember(user_id)
     if (isUserMember) {
         return res.status(400).json({ error: "User is already a member." });
     }
 
+    const canManage = group.canManage(req.user._id)
+    if (!canManage) {
+        return res.status(400).json({ error: "User has no privileges." });
+    }
+
     try {
-        group.addMember(user_id)
-        res.status(200).json(await Group.find({_id: id }))
+        await group.addMember(user_id)
+        res.status(200).json(true)
     } catch (error) {
         res.status(400).json({error: error.message})
     }
@@ -334,9 +427,14 @@ const rejectUser = async (req, res) => {
         return res.status(400).json({ error: "User is not in the pending list." });
     }
 
-    const isUserMember = group.membersID.some(member => member.equals(user_id))
+    const isUserMember = await group.isMember(user_id)
     if (isUserMember) {
         return res.status(400).json({ error: "User is already a member." });
+    }
+
+    const canManage = group.canManage(req.user._id)
+    if (!canManage) {
+        return res.status(400).json({ error: "User has no privileges." });
     }
 
     try {
@@ -344,7 +442,7 @@ const rejectUser = async (req, res) => {
             { _id: id},
             { $pull: { requestID: user_id}}
         )
-        res.status(200).json(await Group.find({_id: id }))
+        res.status(200).json(true)
     } catch (error) {
         res.status(400).json({error: error.message})
     }
@@ -395,6 +493,72 @@ const plantTree = async (req, res) => {
     }
 }
 
+const setPrivacy = async (req, res) => {
+    const user_id = req.user._id
+    const { id } = req.params
+    const { isPrivate } = req.body
+
+    try {
+        let group = await Group.findOne({
+            _id: id,
+        })
+
+        if (!group) {
+            return res.status(404).json("Group not found")
+        }
+
+        const canManage = group.canManage(req.user._id)
+        if (!canManage) {
+            return res.status(400).json({ error: "User has no privileges." });
+        }
+        
+        group.isPrivate = isPrivate === true
+        await group.save()
+        return res.status(200).json(group.isPrivate)
+    } catch (error) {
+        res.status(400).json({error: error.message})
+    }
+}
+
+const setName = async (req, res) => {
+    const user_id = req.user._id
+    const { id } = req.params
+    const { name } = req.body
+
+    try {
+        let group = await Group.findOne({
+            _id: id,
+        })
+
+        if (!group) {
+            return res.status(404).json("Group not found")
+        }
+
+        const canManage = group.canManage(req.user._id)
+        if (!canManage) {
+            return res.status(400).json({ error: "User has no privileges." });
+        }
+
+        if (!validator.isAlphanumeric(name)) {
+            return res.status(400).json({ error: 'Group name can only contain contains only letters and numbers (a-z A-Z 0-9)' });
+        }
+        if (!validator.isLength(name, { min: 4, max: 20 })) {
+            return res.status(400).json({ error: 'Group name must be 4-20 characters long' });
+        }
+        
+        const exist = await Group.findOne({ name: name })
+        if (exist) {
+            return res.status(400).json({ error: "Group name already in use" });
+        }
+
+        group.name = name
+        await group.save()
+        return res.status(200).json(group.name)
+    } catch (error) {
+        res.status(400).json({error: error.message})
+    }
+}
+
 module.exports = {
     createGroup,
     addUser,
@@ -403,12 +567,17 @@ module.exports = {
     getInvite,
     rejectGroup,
     removeMember,
+    leaveGroup,
     deleteGroup,
     searchGroup,
-    joinGroup, 
+    joinGroup,
+    cancelRequest,
     getRequest,
     acceptUser,
     rejectUser,
     getGroupData,
+    revokeInvite,
     plantTree,
+    setPrivacy,
+    setName,
 }
